requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.1.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
psycopg2==2.9.3
PyJWT==2.8.0
python-dotenv==0.20.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
xmltodict==0.13.0
